Refactored useful checker macros into a separate header file whose inclusion can then be made from non .cu files too
