+++ ml-prims/test/cuda_utils.cu
@@ -22,7 +22,7 @@ namespace MLCommon {
 
 TEST(Utils, Assert) {
   ASSERT_NO_THROW(ASSERT(1 == 1, "Should not assert!"));
-  ASSERT_THROW(ASSERT(1 != 1, "Should assert!"), std::runtime_error);
+  ASSERT_THROW(ASSERT(1 != 1, "Should assert!"), Exception);
 }
 
 TEST(Utils, CudaCheck) { ASSERT_NO_THROW(CUDA_CHECK(cudaFree(nullptr))); }
